+++ testing.go
@@ -2,9 +2,7 @@ package spruce
 
 import (
 	"log/slog"
-	"strings"
 	"testing"
-	"time"
 )
 
 // TestingT is the subset of [testing.TB] that is used to write logs.
@@ -15,25 +13,20 @@ type TestingT interface {
 var _ TestingT = (testing.TB)(nil)
 
 // NewTestLogger returns a [slog.Logger] that writes to t.
-func NewTestLogger(t TestingT) *slog.Logger {
-	return slog.New(NewTestHandler(t))
+func NewTestLogger(t TestingT, options ...Option) *slog.Logger {
+	return slog.New(NewTestHandler(t, options...))
 }
 
 // NewTestHandler returns a new [slog.Handler] that writes to t.
-func NewTestHandler(t TestingT) slog.Handler {
-	epoch := time.Now()
-
-	return &handler{
+func NewTestHandler(t TestingT, options ...Option) slog.Handler {
+	h := &handler{
 		log: func(s string) error {
 			t.Log(s)
 			return nil
 		},
-		writeTime: func(w *strings.Builder, rec slog.Record) {
-			elapsed := rec.Time.Sub(epoch)
-			if elapsed > 0 {
-				w.WriteByte('+')
-			}
-			w.WriteString(elapsed.String())
-		},
 	}
+
+	applyOptions(h, options)
+
+	return h
 }